Simplify the api to draw ellipse in bitmap helper
